Switch to TickBasedAnimator and rename a folder
- Replaced AnimationPlayer nodes with TickBasedAnimator nodes.
To make it no longer needed to drag and drop a script to every AnimationPlayer to make it work based on 240 TPS instead of delta time, I created a class that extends AnimationPlayer and gets this exact functionality. It can easily be added to the scene like a built-in node.

- Renamed the "objects" folder to "game_objects" for clarily, with the way RefCounted objects also exist and can be confused with actual scenes.