Add primitive camera control
Point of interest: level.gd
The game constantly checks if the player's center is in contact with the bounding box of any area that isn't the current one.

I attempted making it more faithful to be original games by only requiring the player to slightly touch the outside of the area, but it's extremely hard and confusing, so here's a placeholder for now.