Overhaul file structure of core
Tried to group files in the "core" folder in a way that related .tscn and .gd files are in the same place.

It used to be quite annoying to navigate the file system with scenes being completely separated from scripts, and this update should ideally fix that.

The "core" folder is now split into autoloads, common_objects, misc_assets, ref_counted and special_objects.