Optimize tile maps
The game only used to know if a tile is taken or not, and it used for loops to figure out terrains on the fly which caused noticeable delays.

I made it so the secret, invisible layer which is edited in the scene editor has terrain data by default and now the game just copies "tile_map_data" and assigns them to the 2 visible layers (outline and fill), and it's blazing fast.