Add new AbstractCollider and Trigger features
AbstractColliders now have collision_entered, collision_exited and all_collisions_exited signals. The first two return the node that entered or exited the collider. AbstractColliders also have a multi_activate boolean (whether or not to activate every frame if an object is touching it) and a groups_to_check array (the node groups to check collisions with).
Events now have a state boolean to allow for cleaner enabling or disabling of states.
Triggers have new color customization and a "Button mode" which automatically deactivates the trigger once all objects it checks collisions with have exited it.