Add basic saving
Added a basic save system which currently only saves deaths, time and the current level. Save files are stored in the default user:// folder as a JSON file. The new save_file.gd script uses a dictionary to convert the save to and from JSON using functions save(), view_save() and load_save(). view_save() is used in the saves menu for viewing the saved data before loading the save.