Turn off trails in scene editor
"GPUParticles2D" nodes are off by default and get turned on inside coins' and enemies' _ready() functions. They were causing a noticeable FPS drop in the pink room when dragging the camera around.

Maybe seeing particles in the scene editor could be desired when previewing patterns, but that's a worry for the future because previewing multiple animation players would probably only become possible with a plugin.