Tweak wall push
When the push is deadly large on one axis but small on the other, the deadly large one kills the player while not giving the push but the smaller one still gives the push.

There's still the problem with walls merging in an incorrect order and optimizing them checeking the closest ones first would probably require quadtrees.