Merge wall push into a single step
The step that stretches walls (all for proper crush death physics) is now meant to add stretched walls instead of returning a modified array of walls. Walls getting processed in a single step (instead of 2 which involved extra player movement) will make it simpler to program corner sliding that picks the most "interesting" wall and tries to pull you to its edges.

Getting sandwiched between walls is a case where overlapping walls are actually wanted, as not having them will miss out on certain physics updates.

If you turn on invincibility and allow a wall to crush you, you might still end up a few pixels away from a wall which I'm trying to stop from happening. I believe what I need to fix here is the way the walls are getting merged. The merging mechanic exists to prevent you from getting stuck on flat walls, but it can have unwanted consequences based on the order in which walls are checked.

Sorry for the incomprehensible code. I'm... iterating on it.